add test and modify test file arrangement
